Implement mbrtoc8(), c8rtomb(), and the char8_t typedef.
This change provides implementations for the mbrtoc8 and c8rtomb
functions adopted for C++20 via WG21 P0482R6 and proposed for C2X
via WG14 N2653.  It also provides the char8_t typedef from N2653.

The mbrtoc8 and c8rtomb functions are declared in uchar.h if
either of the C++20 __cpp_char8_t feature test macro or the
_GNU_SOURCE macro are defined.

The char8_t typedef is declared in uchar.h if _GNU_SOURCE is
defined and __cpp_char8_t is not defined (if __cpp_char8_t is
defined, then char8_t is a builtin type).